+++ lib/findUsers.js
@@ -1,6 +1,6 @@
-const Admin = require("../models/admin");
-const Students = require("../models/studentModel");
-const Tutors = require("../models/TutorModel");
+const Admin = require("../models/admin.model");
+const Students = require("../models/student.model");
+const Tutors = require("../models/tutor.model");
 
 const userExist = async (_email, User) => {
   const user = await User.findOne({ email: _email });
