+++ lib/mailingList.js
@@ -1,5 +1,5 @@
 const bcrypt = require("bcryptjs");
-const UserOTP = require("../models/otpModel");
+const UserOTP = require("../models/otp.model");
 
 const { generateRandomNum, addHours } = require("./mathHelpers");
 let { verifyTransPorter, sendMail } = require("../utils/mailing");
