+++ routes/mailing.router.js
@@ -4,7 +4,7 @@ const router = express.Router();
 const {
   handleContactUs,
   handleOTP,
-} = require("../controllers/mailingController");
+} = require("../controllers/mailing.controller");
 
 router.post("/contactUs", handleContactUs);
 router.post("/otp", handleOTP);
